Move some pieces of Dataflow translator to top level
